# hawk-daemon/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "hawk-daemon | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:27371/"
    }
  ],
  "paths": {
...
    "/api/v1/category/create": {
      "post": {
        "tags": [
          "category"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOfObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/category/update": {
      "post": {
        "tags": [
          "category"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOfObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/category/delete": {
      "post": {
        "tags": [
          "category"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOfObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tag/create": {
      "post": {
        "tags": [
          "tag"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOfObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tag/update": {
      "post": {
        "tags": [
          "tag"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOfObject"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tag/delete": {
      "post": {
        "tags": [
          "tag"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOfObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryCreateRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "CategoryNameRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "CategoryUpdateRequest": {
        "required": [
          "name",
          "new_name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "new_name": {
            "type": "string"
          }
        }
      },
...
      "EnvelopeOfObject": {
        "required": [
          "status",
          "data"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "data": { }
        }
      },
...
      "TagCreateRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "TagNameRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "TagUpdateRequest": {
        "required": [
          "name",
          "new_name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "new_name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}